requirements.txt
Django>=4.2
djangorestframework
python-dotenv
dnspython
python-whois
python-nmap
shodan